코드 정리 / 병상배정 타임라인 수정 진행중..

+++ src/main/kotlin/org/sbas/dtos/BaseCodeDto.kt
@@ -15,7 +15,7 @@ data class BaseCodeSaveReq(
     var cdGrpNm: String,
     var cdId: String,
     var cdNm: String,
-    var cdSeq: Long,
+    var cdSeq: Int,
 )
 
 fun BaseCodeSaveReq.toCdIdEntity(): BaseCode {
@@ -38,7 +38,7 @@ data class BaseCodeUpdateReq(
     var cdId: String,
     var cdNm: String,
     var cdVal: String,
-    var cdSeq: Long,
+    var cdSeq: Int,
     var rmk: String,
 )
